Refactored common propagation code into PropagationUtil
git-svn-id: http://encog-java.googlecode.com/svn/trunk/encog-core@1009 f90f6e9a-ac51-0410-b353-d1b83c6f6923

+++ src/org/encog/neural/data/basic/BasicNeuralDataSet.java
@@ -330,9 +330,9 @@ public class BasicNeuralDataSet implements NeuralDataSet, EncogPersistedObject,
 	}
 
 	@Override
-	public void getRecord(int index, NeuralDataPair pair) {
+	public void getRecord(long index, NeuralDataPair pair) {
 		
-		NeuralDataPair source = this.data.get(index);
+		NeuralDataPair source = this.data.get((int)index);
 		pair.getInput().setData(source.getInput().getData());
 		if( pair.getIdeal()!=null ) {
 			pair.getIdeal().setData(source.getIdeal().getData());
